分次检测

+++ src/auto_cut.py
@@ -94,7 +94,7 @@ def cut_video(clip, accept_infos, args):
             if t >= args.max_time:
                 break
             t += args.gap_time
-            index = t * clip.fps
+            index = int(t * clip.fps)
 
     cut_times = get_index_range(accept_infos, args.accept_min_time * clip.fps)
     # 如果任何时间差小于accept_min_time的帧都为False，那中间的部分就不用检测了，直接设置为False
